+++ CHANGELOG.md
@@ -10,7 +10,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added Parser module for parsing PEM public and private keys
 
-## [0.2.0] - 2024-08-01
+- Added function `GenerateSaltAndKey` combining `GenerateSalt` and `GenerateKey` functions into one
+
+## [0.2.2] - 2024-11-10
+
+- Security patch: Updated dependencies
+
+## [0.2.1] - 2024-08-02
+
+### Added
+
+- Added functions to enable use of default `FcryptKey`
+
+## [0.2.0] - 2024-08-02
 
 ### Added
 
@@ -27,6 +39,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Consolidated hashing functions to accept a `hash.Hash` parameter, making the API more flexible and versatile.
 - Generalized `Hash` function for hashing `io.Reader` data using a provided `hash.Hash`.
 
+### Fixed
+
+- Fixed `StreamEncrypt` and `StreamDecrypt` CipherIV issues.
 
 ## [0.1.0] - 2024-07-31